First semi-working fluid-settings.
Fixed the listening capability, we do not listen to the "service" in the messsage, that's the name of the daemon!
Prevent timeout when listening in the CLI (--watch).
Added message to listen to fluid value changes in the CLI (FLUID_SETTINGS_REGISTERED).
Make sure to send the current value on a first listen registration.
Fixed the filename in the first call to the f_settings.load() function.
Fixed the load() so it accepts the '::' scoping parameter.
Transformed the "save-timeout" parameter in an advgetopt duration.
Fixed the time out and made it a value in microseconds instead of seconds.
Fixed the handling of the parameter names. In multiple places, I had to fix the '-'/'_' issue.
Fixed the message sent on a value change to FLUID_SETTINGS_VALUE_UPDATED and added the name parameter.
Moved the VALUE_CHANGED unserialization to the library (since the serialization is there).
Added the MULTIPLE flag on --put, --set (name + value).
Fixed the CLI service name to use its PID so it is unique.
The set_value() and get_value() now return a result which let us know what happened with more details.
Properly check that --list-all was used.
Fixed the handling of the various list options.
Removed the version.h/cpp from the daemon, we use the one in the library instead.
In msg_forget(), the server/service names have to use the get_sent_from_...() functions.
Added some Validator() to command line parameters.
Made the fluid-settings-gui optional (if you do not have Qt available).
Enhanced various error messages/handling.