Updated to work with the new dispatcher.
Added a configuration file for the fluid-settings-cli so we can define the communicator daemon IP:port or Unix URL.
Converted the fluid-settings-cli to make use of a client messenger defined in the library which itself derives from the communicatord::communicator class.
Made the watch work with the library. In most cases, you really won't have much to do in code, just proper definitions in your advgetopt::getopt table.
Fixed the various CLI --list commands. They work now.
Added a --get-default option to the CLI.
Fixed the default communicatord_listen=... parameter in the configuration file.
Fixed the field name, any '_' must be changed to '-' because it gets saved that way in advgetopt.