Moved the sending of the SERVICESTATUS message to the library, it is required by all, not just the CLI.
Added proper support for the automatic watch initialization in the CLI.
Added a function for when we receive the FLUID_SETTINGS_STATUS_READY event. That way we can verify that feature works as expected.
That message includes the total number of errors encountered while processing the messages listened on (if 0, then the "errcnt" field is not present in the message).
Check whether an option in f_opts is already defined, if so, then don't listen on it through the fluid-settings (i.e. you can override the fluid-settings with the command line, .conf, etc.)