Added support for STATUS callbacks.
We cannot have multiple listeners to the same message (once we found a match we stop searching). So here I added a callback manager for those interested in the STATUS message.
Fixed #include filename. I renamed the timespec_ex file to timespec_ex.h to make it easier to remember.