Fixed the msg_status(), send the READY signal when there are no watches.
Removed the callback for the STATUS command. Instead, we now have the
CALLBACK capability of the dispatcher since I added a priority so we
can easily sort the callbacks as expected. No need for the specialized
callbacks.

Removed the "up" value from the names.an, make use of the communicatord one.