Added support for the timeout.
I had a timer implemented, but not yet plugged in. I think that it should
be possible to track all the requests instead of only the last one. That
being said, it may not be useful to get so many time outs. Just one is
probably sufficient to generate an error and know what's going on (or
rather not going on).